backend/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
django-environ==0.4.5
psycopg2-binary==2.9.9
setuptools==69.5.1
sqlparse==0.4.4
uWSGI==2.0.25.1
Pillow==9.0.0
wheel==0.43.0
django-debug-toolbar==4.4.6
stripe==10.8.0
djangorestframework==3.15.2